Reduce footprint of openapi fix
